requirements.txt
fastapi==0.115.5
uvicorn[standard]
pdfplumber==0.11.4
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
httpx==0.27.2
pgvector==0.4.1
Django==4.2.11
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
PyYAML>=6.0
